feat: remove old get balances handlers
